agents/model_summary/pyproject.toml
[project]
name = "model-researcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.5"
dependencies = [
    "huggingface-hub>=0.33.4",
    "fast-agent-mcp>=0.2.45"
]
